+++ include/aedis/resp3/detail/exec.hpp
@@ -132,10 +132,8 @@ struct exec_with_timeout_op {
             case 0: self.complete(ec1, n); break;
             case 1:
             {
-               if (ec2)
-                  self.complete({}, n);
-               else
-                  self.complete(aedis::error::exec_timeout, 0);
+               BOOST_ASSERT_MSG(!ec2, "exec_with_timeout_op: Unexpected completion order.");
+               self.complete(aedis::error::exec_timeout, 0);
 
             } break;